add a new file to harden code, and move function to escape strings
